refactor: refactor and improve code structure
