+++ packages/sn-filter-pane/src/components/ListboxGrid/grid-constants.ts
@@ -1,0 +1,8 @@
+export const COLLAPSED_HEIGHT = 34;
+export const BUTTON_HEIGHT = 34;
+export const ITEM_SPACING = 8;
+export const COLUMN_MIN_WIDTH = 160;
+export const COLUMN_SPACING = 16;
+export const EXPANDED_HEADER_HEIGHT = 48;
+export const SINGLE_GRID_ROW_HEIGHT = 32;
+export const DEFAULT_CSS_HEIGHT = 'calc(100% - 2px)';
